BOCM-51: moved code related to processing of api data from LPA controller to service
